Replaced some floats with doubles

+++ src/hitable_list.cpp
@@ -1,6 +1,6 @@
 #include "./hitable_list.h"
 
-bool HitableList::hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const
+bool HitableList::hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const
 {
     HitRecord temp_rec;
     bool hit_anything = false;
